[#26] refactor : feature/26
전반적인 패키지 구조 리팩토링 했습니다.
