+++ src/main/java/com/health/healther/domain/repository/ReviewRepository.java
@@ -1,8 +1,8 @@
-package com.health.healther.review.domain.repository;
+package com.health.healther.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.health.healther.review.domain.model.Review;
+import com.health.healther.domain.model.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
