+++ src/main/java/com/health/healther/exception/review/NoFoundMemberException.java
@@ -1,4 +1,4 @@
-package com.health.healther.review.exception.review;
+package com.health.healther.exception.review;
 
 public class NoFoundMemberException extends RuntimeException{
     public NoFoundMemberException(String message) {
