+++ src/main/java/com/health/healther/exception/review/NoFoundSpaceException.java
@@ -1,4 +1,4 @@
-package com.health.healther.review.exception.review;
+package com.health.healther.exception.review;
 
 public class NoFoundSpaceException extends RuntimeException{
