+++ src/main/java/com/health/healther/exception/review/ReviewExceptionHandler.java
@@ -1,6 +1,9 @@
-package com.health.healther.review.exception.review;
+package com.health.healther.exception.review;
 
 
+import com.health.healther.exception.review.NoFoundMemberException;
+import com.health.healther.exception.review.NoFoundReviewException;
+import com.health.healther.exception.review.NoFoundSpaceException;
 import com.health.healther.review.exception.ErrorMessage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
